+++ src/main/java/com/svc/ventas/models/dao/ProveedorRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.svc.ventas.models.entity.Proveedor;
 
-public interface ProveedorRepo extends JpaRepository<Proveedor, Integer> {
+public interface ProveedorRepo extends JpaRepository<Proveedor, Long> {
 	
 	List<Proveedor> findAll();
